move some function to checkpoint.go
